# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tools
    description: Tool management operations
  - name: locations
    description: Location management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List all tools
      parameters:
        - name: locationId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
    post:
      operationId: createTool
      tags: [tools]
      summary: Create a new tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolBody"
      responses:
        "201":
          description: Tool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"

  /tools/{id}:
    get:
      operationId: getTool
      tags: [tools]
      summary: Get a tool by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tool details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTool
      tags: [tools]
      summary: Update a tool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateToolBody"
      responses:
        "200":
          description: Tool updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTool
      tags: [tools]
      summary: Delete a tool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tool deleted
        "404":
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Location deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [tools]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/by-location:
    get:
      operationId: getToolsByLocation
      tags: [tools]
      summary: Get tool counts grouped by location
      responses:
        "200":
          description: Tools grouped by location
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationToolCount"

  /dashboard/recent:
    get:
      operationId: getRecentTools
      tags: [tools]
      summary: Get recently added tools
      responses:
        "200":
          description: Recently added tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Tool:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        locationId:
          type: integer
        locationName:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - condition
        - locationId
        - createdAt
        - updatedAt

    CreateToolBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        locationId:
          type: integer
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required:
        - name
        - condition
        - locationId

    UpdateToolBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        locationId:
          type: integer
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true

    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt

    CreateLocationBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required:
        - name

    DashboardSummary:
      type: object
      properties:
        totalTools:
          type: integer
        totalLocations:
          type: integer
        toolsByCondition:
          type: object
          additionalProperties:
            type: integer
        mostActiveLocation:
          type: string
          nullable: true
        recentlyAdded:
          type: integer
      required:
        - totalTools
        - totalLocations
        - toolsByCondition
        - recentlyAdded

    LocationToolCount:
      type: object
      properties:
        locationId:
          type: integer
        locationName:
          type: string
        toolCount:
          type: integer
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required:
        - locationId
        - locationName
        - toolCount

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error