# lib/api-zod/src/generated/types/createLocationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLocationBody {
  name: string;
  address?: string | null;
  latitude?: number | null;
  longitude?: number | null;
}