# lib/api-zod/src/generated/types/createToolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateToolBodyCondition } from "./createToolBodyCondition";

export interface CreateToolBody {
  name: string;
  description?: string | null;
  category?: string | null;
  condition: CreateToolBodyCondition;
  locationId: number;
  latitude?: number | null;
  longitude?: number | null;
}