# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryToolsByCondition } from "./dashboardSummaryToolsByCondition";

export interface DashboardSummary {
  totalTools: number;
  totalLocations: number;
  toolsByCondition: DashboardSummaryToolsByCondition;
  mostActiveLocation?: string | null;
  recentlyAdded: number;
}