# lib/api-zod/src/generated/types/locationToolCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LocationToolCount {
  locationId: number;
  locationName: string;
  toolCount: number;
  latitude?: number | null;
  longitude?: number | null;
}