# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ToolCondition } from "./toolCondition";

export interface Tool {
  id: number;
  name: string;
  description?: string | null;
  category?: string | null;
  condition: ToolCondition;
  locationId: number;
  locationName?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  createdAt: Date;
  updatedAt: Date;
}