# lib/api-zod/src/generated/types/updateToolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateToolBodyCondition } from "./updateToolBodyCondition";

export interface UpdateToolBody {
  name?: string;
  description?: string | null;
  category?: string | null;
  condition?: UpdateToolBodyCondition;
  locationId?: number;
  latitude?: number | null;
  longitude?: number | null;
}